third_party/WebKit/LayoutTests/media/video-dom-autoplay.html

<!DOCTYPE html>
<title>Test media "autoplay" attribute set via DOM.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_false(video.autoplay);
    video.autoplay = true;
    assert_true(video.autoplay);
    assert_not_equals(video.getAttribute("autoplay"), null);

    video.onplay = t.step_func_done(function() {
        assert_false(video.paused);
    });

    video.src = "content/test.ogv";
});
</script>
